+++ src/test/java/org/scify/jedai/configuration/ConfigureRandomlyEndToEndWorkflowCCER.java
@@ -102,7 +102,7 @@ public class ConfigureRandomlyEndToEndWorkflowCCER {
             int bestIteration = 0;
             float bestFMeasure = 0;
             for (int j = 0; j < NO_OF_TRIALS; j++) {
-                float time1 = System.currentTimeMillis();
+                long time1 = System.currentTimeMillis();
 
                 bb.setNextRandomConfiguration();
                 final List<AbstractBlock> blocks = bb.getBlocks(profiles1, profiles2);
@@ -131,7 +131,7 @@ public class ConfigureRandomlyEndToEndWorkflowCCER {
                 ec.setNextRandomConfiguration();
                 final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-                float time2 = System.currentTimeMillis();
+                long time2 = System.currentTimeMillis();
 
                 final StringBuilder matchingWorkflowConf = new StringBuilder();
                 matchingWorkflowConf.append(bb.getMethodConfiguration());
@@ -155,7 +155,7 @@ public class ConfigureRandomlyEndToEndWorkflowCCER {
             System.out.println("\nBest Iteration\t:\t" + bestIteration);
             System.out.println("Best FMeasure\t:\t" + bestFMeasure);
 
-            float time1 = System.currentTimeMillis();
+            long time1 = System.currentTimeMillis();
 
             bb.setNumberedRandomConfiguration(bestIteration);
             final List<AbstractBlock> blocks = bb.getBlocks(profiles1, profiles2);
@@ -175,7 +175,7 @@ public class ConfigureRandomlyEndToEndWorkflowCCER {
             ec.setNumberedRandomConfiguration(bestIteration);
             final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-            float time2 = System.currentTimeMillis();
+            long time2 = System.currentTimeMillis();
 
             final StringBuilder matchingWorkflowConf = new StringBuilder();
             matchingWorkflowConf.append(bb.getMethodConfiguration());
