+++ src/test/java/org/scify/jedai/configuration/StepByStepRandomConfigurationDER.java
@@ -49,8 +49,8 @@ public class StepByStepRandomConfigurationDER {
 
     private final static int NO_OF_TRIALS = 100;
 
-    static int getTotalComparisons(List<AbstractBlock> blocks) {
-        int originalComparisons = 0;
+    static long getTotalComparisons(List<AbstractBlock> blocks) {
+        long originalComparisons = 0;
         for (AbstractBlock block : blocks) {
             originalComparisons += block.getNoOfComparisons();
         }
@@ -95,7 +95,7 @@ public class StepByStepRandomConfigurationDER {
             // local optimization of Block Building
             float bestA = 0;
             int bestIteration = 0;
-            int originalComparisons = profiles.size() * (profiles.size() - 1) / 2;
+            long originalComparisons = profiles.size() * (profiles.size() - 1) / 2;
             for (int j = 0; j < NO_OF_TRIALS; j++) {
                 bb.setNextRandomConfiguration();
                 final List<AbstractBlock> originalBlocks = bb.getBlocks(profiles);
@@ -233,7 +233,7 @@ public class StepByStepRandomConfigurationDER {
             System.out.println("\nBest Iteration\t:\t" + bestIteration);
             System.out.println("Best FMeasure\t:\t" + bestFMeasure);
 
-            float time1 = System.currentTimeMillis();
+            long time1 = System.currentTimeMillis();
 
             em.setNumberedRandomConfiguration(bestIteration);
             final SimilarityPairs sims = em.executeComparisons(finalBlocks);
@@ -241,7 +241,7 @@ public class StepByStepRandomConfigurationDER {
             ec.setNumberedRandomConfiguration(bestIteration);
             final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-            float time2 = System.currentTimeMillis();
+            long time2 = System.currentTimeMillis();
 
             final StringBuilder matchingWorkflowConf = new StringBuilder();
             matchingWorkflowConf.append(bb.getMethodConfiguration());
